refactor(cmdr-index): the writer `entries` test suite splits into eight themed files, so a test is found by its subject instead of by scrolling 2,265 lines
`entries/tests.rs` was the longest file left on the `file-length` list. It becomes `entries/tests/`, following the `copy_tests/` model: `mod.rs` carries the shared imports, the `insert_dir_with_stats` / `insert_file` / `seed_row` fixtures, and the child declarations, and every child opens with `use super::*;`.

The seams follow the behavior each test drives: `insert_upsert.rs` (212), `delete.rs` (141), `delta_propagation.rs` (355), `hardlink_dedup.rs` (420), `symlinks.rs` (257), `move_entry.rs` (587), `bulk_reconcile.rs` (154), `id_counter.rs` (96), plus `mod.rs` (80). Delete-by-id and delete-subtree sit with the other delete paths while their `*_auto_propagates_delta*` twins sit with the upsert propagation tests, because the propagation contract is the thing under test there, not the delete. The bulk-reconcile suppression guard and the id-counter healing tests, both about the writer's own bookkeeping rather than a single mutation, get one small file each rather than a shared grab bag.

Pure move: all 33 tests keep their names and their bodies byte-for-byte, verified by comparing every extracted body against the original; the only line-level differences across the whole split are the six section-banner comments becoming per-file `//!` docs, the eight `mod` declarations, and the eight `use super::*;` lines. The three helpers moved once each into `mod.rs` and are shared, not copied. No production code changed.

Proof: the sorted leaf test-name list from `cargo nextest list -p cmdr-index -- writer::entries` is identical before and after (33 names), and `cargo nextest run` reports 33 passed. `file-length` shrink-wrapped the stale 2,138-line entry away and added none; `jscpd-rust` is unchanged; `clippy`, `rustfmt`, `lock-poison`, `error-string-match`, `test-sleep`, and `probe-unwrap-justified` are green.
